Session juin 2026: NB + time-weighted params + Shin + Elo + θ-MLE + Gemma4
- Negative Binomial remplace Poisson dans StatisticalEngine.js
  (θ auto-adapté par ligue, _getThetaForLeague)
- Time-weighted form + attack/defense (exp decay λ=0.008/jour)
- Elo rating system (services/eloRatingService.js, K=32, HA=100)
- Shin method comme défaut dans ValueBetEngine.js (via quantService)
- Auto-θ tuning par MLE (services/thetaOptimizer.js, méthode des moments)
- Gemma4/DeepSeek tactical briefings (services/gemma4Service.js)
- Sofascore xG service (services/sofascoreXgService.js)
- Nouveaux endpoints: /api/theta/optimize, /api/elo, /api/elo/update
- /health amélioré (dist NB, nb Elo, nb θ)
- Mode Chirurgical dans Dashboard (Conf≥75%, Prob≥55%)
- Rétabli computeFormFromHistory avec historique SQLite
- Render: healthCheckPath, notifications email/slack, env groups
- Frontend: plus de derivePreciseFTScore (backend score autoritaire)
- Cron: AutoHeal */15min, enrichissement 6x/jour, lookahead 3j
- Jest: downgrade 29.7.0, modulePathIgnorePatterns pour .kilo

+++ core/database.js
@@ -263,7 +263,7 @@ function initSchema() {
                 name       TEXT UNIQUE NOT NULL,
                 normalized TEXT NOT NULL,
                 league     TEXT,
-                last_seen  INTEGER
+                last_seen  BIGINT
             );
             CREATE INDEX IF NOT EXISTS idx_team_registry_normalized ON team_registry(normalized);
